refactor: rename model to domain
