+++ mostlyai/client/connectors.py
@@ -1,7 +1,7 @@
 from typing import Any, Iterator, Optional, List, Dict, Union
 
 from mostlyai.client.base import DELETE, GET, PATCH, POST, Paginator, _MostlyBaseClient
-from mostlyai.client.model import (
+from mostlyai.client.domain import (
     Connector,
     ConnectorListItem,
     SyntheticDatasetConfig,
