+++ mostlyai/client/generators.py
@@ -5,7 +5,7 @@ import re
 import pandas as pd
 
 from mostlyai.client.base import DELETE, GET, PATCH, POST, Paginator, _MostlyBaseClient
-from mostlyai.client.model import (
+from mostlyai.client.domain import (
     Generator,
     JobProgress,
     GeneratorListItem,
