+++ tests/end_to_end/end_to_end.py
@@ -1,7 +1,7 @@
 from unittest import skip
 
 from mostlyai import MostlyAI
-from mostlyai.client.model import GeneratorConfig, SyntheticDatasetConfig
+from mostlyai.client.domain import GeneratorConfig, SyntheticDatasetConfig
 import pandas as pd
