+++ tools/README.md
@@ -3,7 +3,7 @@
 ## model extension
 
 While `datamodel-codegen` is a very handy tool for converting OpenAPI definitions into
-pydantic objects (see `mostlyai/client/model.py`), it lacks the direct ability of adding extra functionality (e.g. methods)
+pydantic objects (see `mostlyai/client/domain.py`), it lacks the direct ability of adding extra functionality (e.g. methods)
 to the classes it creates. For example, having `Generator.add_table(...)` is out of its scope, but
 there's a trick being used here:
 - `datamodel-codegen` works with `Jinja2` templates, and custom templates can be specified
@@ -16,7 +16,7 @@ Moreover, all of that happens by simply running `make gen-public-model`
 
 ## Updating public model (based on `public-api.yaml`)
 
-`make gen-public-model` does all the required actions to rewrite and format `mostlyai/model.py`. The relevant code sections from `tools/model.py` are stitched into the resulted `mostlyai/model.py`
+`make gen-public-model` does all the required actions to rewrite and format `mostlyai/client/domain.py`. The relevant code sections from `tools/model.py` are stitched into the resulted `mostlyai/client/domain.py`
 
 # Release Workflow
 
@@ -24,7 +24,7 @@ The release workflow itself is made easy using the `Makefile`.
 
 ```bash
 BUMP_TYPE=patch make update-version-gh
-# create PR, squash&merge PR, create release on GitHub with the new tag 
+# create PR, squash&merge PR, create release on GitHub with the new tag
 make release-pypi
 ```
