+++ tools/postproc_model.py
@@ -1,5 +1,5 @@
 # Constant for the file path
-FILE_PATH = "mostlyai/client/model.py"
+FILE_PATH = "mostlyai/client/domain.py"
 
 # Dictionary for enum replacements
 enum_replace_dict = {
